change RR into priority-based scheduling
Add two structs for priority scheduling.
one record all tasks which have createdthe other record the data of tasks,
inorder to let the scheduler to choose the task which has heighest priority.
when the task is creating, os will record the priority of task and insert into ReadyItemList.
when the system tick is happended, it will choose the task having priority to excute.

Add "-std=c99" to CFLAGS in Makefile, inorder to for loop inline.